requirements.txt
flask==2.0.1
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
seaborn==0.11.2
plotly==5.3.1
kagglehub
pillow==8.3.2